+++ build.py
@@ -1,8 +1,8 @@
 """Helper script for building a site; used in Docker build process."""
 
 from arxiv.base.globals import get_application_config
-from markdown.factory import create_web_app
-from markdown.build import build_site
+from marxdown.factory import create_web_app
+from marxdown.build import build_site
 
 if __name__ == '__main__':
     app = create_web_app()
